Split io/filepath.go into three files, move packages into pkg folder
